+++ NEWS.md
@@ -6,8 +6,8 @@ Version 0.11.1 - dev
 --------------------
 
 - NEW: `Meshbuilder.from_polyface()` to interface to `POLYFACE` and `POLYMESH` 
-- NEW: `Meshbuilder.render_polyface()` create `POLYFACE` objects 
-- BUGFIX: `add_polymesh()` and `add_polyface()` - entity casting before adding to layout
+- NEW: `Meshbuilder.render_polyface()` create `POLYFACE` objects
+- CHANGES: refactor Auditor() into a DXF document fixer, fixes will be applied automatically (work in progress)
 
 Version 0.11 - 2020-02-15
 -------------------------