+++ NEWS.md
@@ -2,14 +2,15 @@
 News
 ====
 
-Version 0.11.1 - dev
---------------------
+Version 0.11.1 - 2020-02-29
+---------------------------
 
 - NEW: `Meshbuilder.from_polyface()` to interface to `POLYFACE` and `POLYMESH` 
 - NEW: `Meshbuilder.render_polyface()` create `POLYFACE` objects
 - NEW: `MeshAverageVertexMerger()` an extended version of `MeshVertexMerger()`, location of merged vertices 
   is the average location of all vertices with the same key
-- NEW: `ezdxf.addons.iterdxf` experimental feature to iterate over modelspace entities of really big DXF files (>5 GB)
+- NEW: `ezdxf.addons.iterdxf` iterate over modelspace entities of really big DXF files (>1 GB) without loading 
+  them into memory
 - NEW: `ezdxf.addons.r12writer` supports POLYFACE and POLYMESH entities
 - NEW: `Layout.add_foreign_entity()` copy/move **simple** entities from another DXF document or add unassigned
   DXF entities to a layout